Add junction table support
each time an order is processed, a new row is added to the junction table with the orderid, itemid, and quantity. This lets you see what items were ordered when. It also uses foreign keys so it is constrained to data that is already in the items, and orders table.